requirements.txt
Pillow
torch
transformers
einops
pyyaml